pyproject.toml
[project]
name = "appo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jupyter>=1.1.1",
    "lightning>=2.5.2",
    "plotly-express>=0.4.1",
    "pydantic>=2.11.7",
    "pytorch-gpu>=0.0.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
    "timm>=1.0.19",
    "torchvision>=0.23.0",
    "uvicorn>=0.35.0",
]
